back to normal

+++ backend/controllers/cards.js
@@ -2,7 +2,7 @@ const Card = require('../models/card');
 
 const DataError = require('../errors/DataError');
 const NotFoundError = require('../errors/NotFoundError');
-// const ForbiddenError = require('../errors/ForbiddenError');
+const ForbiddenError = require('../errors/ForbiddenError');
 
 const createCard = (req, res, next) => {
   const { name, link } = req.body;
@@ -33,9 +33,9 @@ const deleteCard = (req, res, next) => {
   Card.findById(req.params.cardId)
     .orFail(() => new NotFoundError('Нет карточки с таким id'))
     .then((card) => {
-      // if (card.owner.toString() !== req.user._id) {
-      //   throw new ForbiddenError('Недостаточно прав для удаления карточки');
-      // }
+      if (card.owner.toString() !== req.user._id) {
+        throw new ForbiddenError('Недостаточно прав для удаления карточки');
+      }
       Card.deleteOne(card).then(() => {
         res.send({ card });
       })
